float better for svg

+++ src/main/java/edu/sc/seis/TauP/AbstractPathSegment.java
@@ -221,7 +221,7 @@ public abstract class AbstractPathSegment {
         if (isDegenerate) {
             TimeDist td = path.get(0);
             double[] xy = SvgEarth.xyForDistDegRadius( td.getDistDeg(), radiusOfEarth - td.getDepth());
-            pw.println("      <circle class=\"degeneratesegment " + getCssClasses() + "\" cx=\""+xy[0] + "\" cy=\""+xy[1] + "\" r=\""+minPolylineSize+"\"/>");
+            pw.println("      <circle class=\"degeneratesegment " + getCssClasses() + "\" cx=\""+((float)xy[0]) + "\" cy=\""+((float)xy[1]) + "\" r=\""+((float)minPolylineSize)+"\"/>");
         } else {
             pw.println("      <polyline class=\"" + getCssClasses() + "\" points=\"");
             String prevLine = "";
